requirements.txt
altair==4.2.2
anyio==3.6.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fire==0.5.0
fonttools==4.39.4
funcy==2.0
future==0.18.3
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
icon-font-to-png==0.4.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jianshuresearchtools==2.11.0
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.1
mdurl==0.1.2
numexpr==2.8.4
numpy==1.24.3
packaging==23.1
palettable==3.3.3
pandas==1.3.4
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prettytable==3.7.0
protobuf==3.20.3
pyarrow==12.0.0
pydeck==0.8.1b0
pyecharts==2.0.3
Pygments==2.15.1
pyLDAvis==3.4.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
streamlit==1.23.1
stylecloud==0.5.2
tenacity==8.2.2
termcolor==2.3.0
threadpoolctl==3.1.0
tinycss==0.4
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
ttkbootstrap==1.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
wordcloud==1.9.1.1
zipp==3.15.0
